Rearrange seeders to not seed Users in Demo
