+++ config/internal-users.php
@@ -1,5 +1,5 @@
 <?php
 
 return [
-    'emails' => explode(',', env('DEMO_INTERNAL_USER_EMAILS') ?? ''),
+    'emails' => env('DEMO_INTERNAL_USER_EMAILS') ? explode(',', env('DEMO_INTERNAL_USER_EMAILS')) : null,
 ];
